Remove Test Header in generators
